🐛 (User login) Fix internal server error when user is not found
We were trying to create a `User` object with data from MongoDB after searching for a user through e-mail. If a user matching that e-mail was not found in MongoDB, we were passing `**None` to the `User(**user_data)` constructor, which resulted in an Internal Server Error.

This is now fixed by flashing a message and redirecting to the login page if no user data was found for the e-mail.